+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [PEP 440](https://www.python.org/dev/peps/pep-0440/)
 and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.2.0]
+
+### Added
+- Added AWS Aurora RDS database for fire extent polygons and metadata.
+- Added the ability to automatically push fire extent polygons and metadata to the Aurora database.
+
 ## [3.1.0]
 
 ### Added